Move some functions to their correct place. Fix some EDA tests.
