+++ niimpy/preprocessing/util.py
@@ -159,110 +159,6 @@ def occurrence(series, bin_width=720, grouping_width=3600):
     gb2.index = gb2.loc[:, ['day', 'hour']].apply(lambda row: pd.Timestamp('%s %s:00'%(row['day'], row['hour'])), axis=1)
     return gb2
 
-def create_timeindex_dataframe(nrows, ncols, random_state=None, freq=None):
-    """Create a datetime index Pandas dataframe 
-    
-    Parameters
-    ----------
-    nrows : int
-        Number of rows
-    ncols : int
-        Number of columns
-    random_state: float, optional
-        Random seed. If not given, default to 33.
-    freq: string, optional:
-        Sampling frequency.
-    Returns
-    -------
-    df : pandas.DataFrame
-        Pandas dataframe containing sample data with random missing rows.    
-    """
-    
-    # Create a nrows x ncols matrix
-    data = np.random.uniform(100, size=(nrows, ncols))
-    df = pd.DataFrame(data)
-    
-    if freq is None:
-        freq='h'
-    idx = _makeDatetimeIndex(nrows, freq=freq)
-    df = df.set_index(idx)
-    
-    return df
-
-
-def create_missing_dataframe(nrows, ncols, density=.9, random_state=None, index_type=None, freq=None):
-    """Create a Pandas dataframe with random missingness.
-    
-    Parameters
-    ----------
-    nrows : int
-        Number of rows
-    ncols : int
-        Number of columns
-    density: float
-        Amount of available data
-    random_state: float, optional
-        Random seed. If not given, default to 33.
-    index_type: float, optional
-        Accepts the following values: "dt" for timestamp, "int" for integer.
-    freq: string, optional:
-        Sampling frequency. This option is only available is index_type is "dt". 
-        
-    Returns
-    -------
-    df : pandas.DataFrame
-        Pandas dataframe containing sample data with random missing rows.    
-    """
-    
-    # Create a nrows x ncols matrix
-    data = np.random.uniform(100, size=(nrows, ncols))
-    df = pd.DataFrame(data)
-    
-    if index_type:
-        if index_type == "dt":
-            if freq is None:
-                freq='h'
-            idx = _makeDatetimeIndex(nrows, freq=freq)
-            df = df.set_index(idx)
-        elif index_type == "int":
-            return
-        else: 
-            raise ValueError("Can't recognize index_type. Try the following values: 'dt', 'int'.")
-            
-    i_idx, j_idx = _create_missing_idx(nrows, ncols, density, random_state)
-    df.values[i_idx, j_idx] = None
-    return df
-
-def _makeDatetimeIndex(k=10, freq='B', name=None):
-    dt = datetime(2022, 1, 1)
-    dr = pd.bdate_range(dt, periods=k, freq=freq, name=name)
-    return pd.DatetimeIndex(dr, name=name)
-
-def _create_missing_idx(nrows, ncols, density, random_state=None):
-    if random_state is None:
-        random_state = np.random
-    else:
-        random_state = np.random.RandomState(random_state)
-
-    # below is cribbed from scipy.sparse
-    size = int(np.round((1 - density) * nrows * ncols))
-    # generate a few more to ensure unique values
-    min_rows = 5
-    fac = 1.02
-    extra_size = min(size + min_rows, fac * size)
-
-    def _gen_unique_rand(rng, _extra_size):
-        ind = rng.rand(int(_extra_size))
-        return np.unique(np.floor(ind * nrows * ncols))[:size]
-
-    ind = _gen_unique_rand(random_state, extra_size)
-    while ind.size < size:
-        extra_size *= 1.05
-        ind = _gen_unique_rand(random_state, extra_size)
-
-    j = np.floor(ind * 1. / nrows).astype(int)
-    i = (ind - j * nrows).astype(int)
-    return i.tolist(), j.tolist()
 
 def aggregate(df, freq, method_numerical='mean', method_categorical='first', groups=['user']):
     """ Grouping and resampling the data. This function performs separated resampling
